refactor: remove mentions of js and pro from filenames
