+++ projects/fingerprintjs-pro-angular/src/lib/fingerprint-angular.service.spec.ts
