+++ projects/fingerprintjs-pro-angular/src/lib/tokens/fingerprint-angular-settings-token.ts
